🧹 Move sort button to own file
